requirements.txt
solana
python-dotenv
requests